fix: Unexpected behaviour when passed an invalid template to init --template (#6)
* wip

* chore: removed chai as dev-dep

* chore: adjusted jest to test properly

* feat: upgrade npm version

+++ src/commands/init/index.ts
@@ -29,6 +29,11 @@ export default class Init extends Command {
 
       this.log(`${MESSAGES.BOOTSTRAPING_APP} ${args.name}`)
 
+      if (!TEMPLATES_URL[flags.template]) {
+        this.log(ERRORS.INVALID_FLAG)
+        return
+      }
+
       exec(`git clone ${TEMPLATES_URL[flags.template]} ${args.name}`)
       exec('rm -rf .git && git init')
